Add all missing files from project

+++ app/src/main/java/com/dbeginc/simpleweather/data/proxies/remote/Wind.kt
@@ -3,6 +3,7 @@ package com.dbeginc.simpleweather.data.proxies.remote
 import com.google.gson.annotations.SerializedName
 
 data class Wind(@SerializedName("deg")
-                val deg: Int = 0,
+                val deg: Double = 0.0,
                 @SerializedName("speed")
-                val speed: Int = 0)
+                val speed: Double = 0.0
+)
